Script for Kamikaze Ship Enemy
It basically targets the player until a certain point and keeps moving forward. This allows the player to get out of the way before the kamikaze ship gets in range to deal damage to the player. If the player dodges the kamikaze ship, the ship will continue on for a bit before destroying itself.